services/service-a/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
httpx==0.26.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-instrumentation-httpx==0.43b0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
prometheus-client==0.19.0
python-json-logger==2.0.7
